requirements.txt
# Runtime dependencies for examples/templates
openai>=1.0.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.2.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
